+++ lib/dradis-appspider.rb
@@ -1,0 +1,8 @@
+# hook to the framework base clases
+require 'dradis-plugins'
+
+# load this add-on's engine
+require 'dradis/plugins/appspider'
+
+# load supporting AppSpider classes
+require 'appspider/vuln'