+++ lib/dradis/plugins/appspider/field_processor.rb
@@ -1,8 +1,8 @@
-module Dradis::Plugins::NTOSpider
+module Dradis::Plugins::Appspider
   class FieldProcessor < Dradis::Plugins::Upload::FieldProcessor
 
     def post_initialize(args={})
-      @nto_object = ::NTOSpider::Vuln.new(data)
+      @appspider_object = ::Appspider::Vuln.new(data)
     end
 
     def value(args={})
@@ -15,7 +15,7 @@ module Dradis::Plugins::NTOSpider
       # The XML uses a <Method> entity, but 'method' is a reserved word here so:
       name = 'vuln_method' if name == 'method'
 
-      @nto_object.try(name) || 'n/a'
+      @appspider_object.try(name) || 'n/a'
     end
   end