+++ lib/dradis/plugins/appspider/importer.rb
@@ -1,15 +1,15 @@
-module Dradis::Plugins::NTOSpider
+module Dradis::Plugins::Appspider
   class Importer < Dradis::Plugins::Upload::Importer
 
     BAD_FILENAME_ERROR_MESSAGE = \
       "The uploaded file should be named VulnerabilitiesSummary.xml. "\
-      "You'll find VulnerabilitiesSummary.xml inside the /report subdirectory in NTO Spider's output."
+      "You'll find VulnerabilitiesSummary.xml inside the /report subdirectory in AppSpider's output."
     NO_VULNSUMMARY_ERROR_MESSAGE = \
       "A proper root element (/VulnSummary) wasn't detected in the uploaded file. "\
-      "Ensure the file you uploaded comes from a NTOSpider report."
+      "Ensure the file you uploaded comes from a AppSpider report."
     NO_VULNS_ERROR_MESSAGE = \
       "No vulnerabilities were detected in the uploaded file (/VulnSummary/VulnList/Vuln). "\
-      "Ensure the file you uploaded comes from a NTOSpider report."
+      "Ensure the file you uploaded comes from a AppSpider report."
 
     # The framework will call this function if the user selects this plugin from
     # the dropdown list and uploads a file.
@@ -40,7 +40,7 @@ module Dradis::Plugins::NTOSpider
       end
 
       @doc.xpath('/VulnSummary/VulnList/Vuln').each do |xml_vuln|
-        vuln = ::NTOSpider::Vuln.new(xml_vuln)
+        vuln = ::Appspider::Vuln.new(xml_vuln)
 
         host_node_label = xml_vuln.at_xpath('./WebSite').text
         host_node_label = URI.parse(host_node_label).host rescue host_node_label
@@ -68,7 +68,7 @@ module Dradis::Plugins::NTOSpider
     private
     def log_error_and_return(message)
       logger.fatal { message }
-      content_service.create_note text: "#[Title]#\nNTO upload error\n\n#[Description]#\n#{ message }"
+      content_service.create_note text: "#[Title]#\nAppSpider upload error\n\n#[Description]#\n#{ message }"
     end
   end
 end